retrieved object by float key
